Split up sections into client components and fix frontend plans page rendering issue
