+++ .env.local.example
@@ -4,6 +4,6 @@ AUTH_GOOGLE_ID=google_client_id_here
 AUTH_GOOGLE_SECRET=google_client_secret_here
 
 AUTH_URL=http://localhost:3000
-BACKEND_API_URL=http://localhost:8080
+NEXT_PUBLIC_BACKEND_API_URL=http://localhost:8080
 
 MONGODB_URI=mongo_db_credentials_here
